# Optimize bulk writing of all blocks of bloom filter