# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Part of #3320.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

```
write_batch primitive/4096 values primitive with bloom filter                                                      
                        time:   [4.2750 ms 4.2905 ms 4.3059 ms]                                                    
                        thrpt:  [40.995 MiB/s 41.143 MiB/s 41.291 MiB/s]                                           
                 change:                                                                                           
                        time:   [-18.394% -16.070% -13.847%] (p = 0.00 < 0.05)                                     
                        thrpt:  [+16.072% +19.147% +22.541%]                              
                        Performance has improved.                                                                  
write_batch primitive/4096 values primitive non-null with bloom filter     
                        time:   [4.4500 ms 4.5848 ms 4.7319 ms]                 
                        thrpt:  [36.582 MiB/s 37.756 MiB/s 38.900 MiB/s]
                 change:                                                                                           
                        time:   [-21.355% -17.148% -12.445%] (p = 0.00 < 0.05)                                     
                        thrpt:  [+14.214% +20.697% +27.153%]                                                       
                        Performance has improved.                                                                  
write_batch primitive/4096 values string with bloom filter
                        time:   [945.94 µs 946.62 µs 947.37 µs]
                        thrpt:  [84.060 MiB/s 84.126 MiB/s 84.186 MiB/s]
                 change:
                        time:   [-26.414% -25.196% -23.997%] (p = 0.00 < 0.05)
                        thrpt:  [+31.573% +33.682% +35.895%]
                        Performance has improved.
write_batch primitive/4096 values string dictionary with bloom filter
                        time:   [494.68 µs 495.13 µs 495.60 µs]
                        thrpt:  [97.137 MiB/s 97.230 MiB/s 97.318 MiB/s]
                 change:
                        time:   [-19.467% -18.637% -17.900%] (p = 0.00 < 0.05)
                        thrpt:  [+21.803% +22.905% +24.173%]
                        Performance has improved.
write_batch primitive/4096 values string non-null with bloom filter
                        time:   [1.0266 ms 1.0275 ms 1.0284 ms]
                        thrpt:  [76.484 MiB/s 76.553 MiB/s 76.618 MiB/s]
                 change:
                        time:   [-28.422% -28.120% -27.852%] (p = 0.00 < 0.05)
                        thrpt:  [+38.604% +39.121% +39.709%]
                        Performance has improved.
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->