@viirya 2022-12-13
Yea, seems other parts of the parquet writer could have benefit from BufWriter if we can add it higher.